create mock auth and validator modules
