+++ dashboard/tests/unit/mockvalidator.ts
@@ -1,0 +1,8 @@
+import APISpec from "./openapi.json";
+import { APIValidator } from "@/types/validation/Validator";
+
+const $validator = new APIValidator();
+
+$validator.getAPISpec = jest.fn().mockResolvedValue(APISpec);
+
+export { $validator };
